exam/gemini-file-search-main/requirements.txt
google-genai==1.49.0
python-dotenv==1.2.1

# Testing dependencies
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1